Add chart of cumulative achievement percentages
Display a sparkline of a users cumulative achievement percentages in their games. Required rolling back the version of jQuery to 1.10.0, but I think I may well get more value from the sparklines in the future.

Allow users to toggle between text and chart, storing the preference in GM's local cache.